Removed cinfo class and renamed mesos config
The mesos config object was "deprecated", but not really.
We still use much of that config, so instead I "undprecated" it and
simply renamed it. (hopefully "backend" is a better name).

Also removed the deprecated Container Info class as it is no longer
used.
