+++ titus-server-master/src/main/java/com/netflix/titus/master/kubernetes/pod/KubePodModule.java
@@ -47,7 +47,7 @@ import com.netflix.titus.master.kubernetes.pod.taint.TaintTolerationFactory;
 import com.netflix.titus.master.kubernetes.pod.topology.DefaultTopologyFactory;
 import com.netflix.titus.master.kubernetes.pod.topology.TopologyFactory;
 import com.netflix.titus.master.kubernetes.pod.v1.V1SpecPodFactory;
-import com.netflix.titus.master.mesos.MesosConfiguration;
+import com.netflix.titus.master.config.BackendConfiguration;
 import com.netflix.titus.master.service.management.ApplicationSlaManagementService;
 
 public class KubePodModule extends AbstractModule {
@@ -64,12 +64,12 @@ public class KubePodModule extends AbstractModule {
     }
 
     /**
-     * TODO Move properties from {@link MesosConfiguration} to other configuration classes.
+     * TODO Move properties from {@link BackendConfiguration} to other configuration classes.
      */
     @Provides
     @Singleton
-    public MesosConfiguration getMesosConfiguration(ConfigProxyFactory factory) {
-        return factory.newProxy(MesosConfiguration.class);
+    public BackendConfiguration getBackendConfiguration(ConfigProxyFactory factory) {
+        return factory.newProxy(BackendConfiguration.class);
     }
 
     @Provides
